+++ src/DDEmodel.jl
@@ -1,5 +1,3 @@
-using DelayDiffEq, LsqFit
-
 """
         This file contains functions to fit the data to a Delay Differential Equation model, and find the parameters
 """
@@ -22,20 +20,18 @@ exp_model(t, p) = @. p[1]*exp(t*p[2]) # exponential model
 
 function find_history(g1::Matrix{Float64}, g2::Matrix{Float64})
     """ to find the history function based on the control trial. """
-    control_g1 = g1[:, 1] # y1
-    control_g2= g2[:, 1] # y2
-    p0 = [1.0, 0.5]
     times = range(0.0; stop = 95.5, length = 192) # x
-    fit_g1 = curve_fit(exp_model, times, control_g1, p0) 
-    fit_g2 = curve_fit(exp_model, times, control_g2, p0) 
+    fit_g1 = curve_fit(exp_model, times, g1[:, 1], [1.0, 0.5]) 
+    fit_g2 = curve_fit(exp_model, times, g2[:, 1], [1.0, 0.5]) 
 
     return coef(fit_g1), coef(fit_g2)
 end
 
+
 # define problem generator (optimization in log space)
 function prob_generator(prob, p)
     exp_p = exp.(p)
-    remake(prob; p = exp_p, constant_lags = [exp_p[3], exp_p[4]])
+    remake(prob; p=exp_p, constant_lags=[exp_p[3], exp_p[4]])
 end
 
 
@@ -86,4 +82,52 @@ function optimization(g1::Matrix{Float64}, g2::Matrix{Float64}, g1_0::Vector{Flo
     results_dde = bboptimize(obj; SearchRange=bound, TraceMode=:silent, MasSteps=num_steps, Method=:adaptive_de_rand_1_bin_radiuslimited)
     new_guess = best_candidate(results_dde)
     return best_fitness(results_dde), exp.(new_guess)
+end
+
+# this function is used for DDE fitting with constant delay
+function DDEsolveConstDelay(g1, g2, g1_0, g2_0, initial_guess, j)
+
+    # initial guess has 34 values, there are 8 * 4 parameters for transition and death rates for each conentration, and the last two are delays for G1 and G2.
+    times = range(0.0; stop = 95.5, length = 192)
+
+    # history function
+    fit1, fit2 = find_history(g1, g2)
+    h(p, t) = [exp_model(t, fit1); exp_model(t, fit2)]
+
+    # the parameters for a specific concentration, j determines that
+    pars = append!([initial_guess[4*j - 3], initial_guess[4*j - 2]], [initial_guess[33], initial_guess[34], initial_guess[4*j - 1], initial_guess[4*j]])
+    # problem
+    prob = DDEProblem(DDEmodel, [g1_0[j], g2_0[j]], h, extrema(times), pars;
+                          constant_lags = [pars[5], pars[6]])
+
+    # the algorithm to solve the dde problem
+    alg = MethodOfSteps(AutoTsit5(Rosenbrock23()); constrained=true)
+
+    solve(prob, alg)
+end
+
+function cost_functionConstDelay(initial_guess, g1, g2, g1_0, g2_0)
+    times = range(0.0; stop = 95.5, length = 192)
+
+    total_cost = zeros(8)
+    for j in 1:8
+        sol = DDEsolveConstDelay(g1, g2, g1_0, g2_0, initial_guess, j)
+        res1 = sum((sol(times, idxs=1).u - g1[:, j]).^2)
+        res2 = sum((sol(times, idxs=2).u - g2[:, j]).^2)
+        total_cost[j] = res1 + res2
+    end
+    # the output of the cost function must be an scalar, so we sum all the squared errors.
+    return sum(total_cost)
+end
+
+function optimization_constantDelay(g1, g2, g1_0, g2_0, initial_guess, lower, upper, num_steps)
+
+    # the cost function must just be a function of parameters
+    cost_fcn(initial_guess) = cost_functionConstDelay(initial_guess, g1, g2, g1_0, g2_0)
+    bound = collect(zip(lower, upper))
+    # optimizing
+    results_dde = bboptimize(cost_fcn; SearchRange=bound, MaxSteps=num_steps, Method=:adaptive_de_rand_1_bin_radiuslimited)
+
+    new_guess = best_candidate(results_dde)
+    return best_fitness(results_dde), new_guess
 end
