+++ src/plot.jl
@@ -1,5 +1,3 @@
-using Plots, CSV, DataFrames, Measures
-include("importData.jl")
 gr()
 """ 
         This file contains a function to plot the parameters against the  drug concentrations, in a 2x2 subplot.
